backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9
google-generativeai>=0.3.0
python-dotenv>=1.0.0
django-cors-headers
dj-database-url>=2.1.0
django-filter>=23.0
djangorestframework-simplejwt
django-allauth
dj-rest-auth[with_social]
